+++ src/app/result/result.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit, Input } from '@angular/core';
 
-// const MAX_SIZE = 5e5;
+const MAX_SIZE = 1e5;
+const ERROR_MAX_OUTPUT_SIZE = 'Error: Output exceeded maximum size allowed';
 
 @Component({
   selector: 'app-result',
@@ -9,22 +10,24 @@ import { Component, OnInit, Input } from '@angular/core';
 })
 export class ResultComponent implements OnInit {
   result: string;
+  visibleResult: string;
 
   constructor() { }
 
   clean() {
     this.result = '';
+    this.visibleResult = '';
   }
 
   addLine(line) {
-    const size = this.result.length;
-    const newSize = size + line.length + 1;
-
-    // if (this.result.length >= MAX_SIZE) {
-      // console.log('oversized', size, newSize);
-    // } else {
     this.result += line + '\n';
-    // }
+
+    if (this.result.length < MAX_SIZE) {
+      this.visibleResult = this.result;
+    } else {
+      this.visibleResult += `${this.result.slice(0, MAX_SIZE)}\n\n${ERROR_MAX_OUTPUT_SIZE}`;
+      return false;
+    }
   }
 
   ngOnInit() {