+++ changelog.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Fixed
 - Throttle data posted from webworker to avoid interface blocking.
+- Limit maximum output size.
 
 ## [1.0.0] - 2020-01-04
 ### Added